+++ presentation/voca/src/main/java/com/hilingual/presentation/voca/model/VocaItemDetail.kt
@@ -5,7 +5,7 @@ import kotlinx.collections.immutable.ImmutableList
 
 @Immutable
 internal data class VocaItemDetail(
-    val phraseId: Int,
+    val phraseId: Long,
     val phrase: String,
     val phraseType: ImmutableList<String>,
     val explanation: String,
